[Resolves PW-614] Monkeypatch for JSONAPI::LinkBuilder
The requests are being redirected to the service based on the NGINX regex matchers, so from the POV of the service that is running the jsonapi-resources, the endpoint is `/` instead of `/<service_name>`. In order to keep the service name in the url, we had to do the monkey patching.

This monkeypatch relies on the class namespace for the resource being rendered.

E.g. for game service

*app/resources/engagement_resource.rb*

```
class EngagementResource < Game::ApplicationResource
...
end
```