[Resolves PW-778] IAM get all policies api crashing
Mainly this addresses the issue with users trying to get the policies under IAM service. Took the chance to write the initial specs for the Policies check and configure IAM to run the specs without issues both on local and CI.

This updates a couple of gems as well:
- nokogiri
- bundler

Removes minitest from the gem spec of sdk as we are using Rspec for our tests.