[PW-1351] Create Rspec Shared Context for Cognito user
[Create Rspec Shared Context for Cognito user](https://perxtechnologies.atlassian.net/browse/PW-1351)

- Created rspec shared context to log in as cognito user 

# How does the implementation addresses the problem

This shared example is needed as when a user posts a new survey answer/game transaction, the model resource tries to grab the `cognito_user_id` which is stored in the JWT token. If the `user_id` isn't here, the response fails and user_id is a compulsory field on the model.