[PW-1819] API necessary to handle reset password
[API necessary to handle reset password](https://perxtechnologies.atlassian.net/browse/PW-1819)

# Describe the changes made. What does this PR changes that might be critical. If any critical decisions have been made, make sure you explain the rationale for these decisions.

- infra:
  - Only filter passwords in production

- iam:
  - Everything should be happening under the Iam:: namespace instead of Devise::
  - add support for changing passwords
  - Rename Device::ApplicationsController to Iam::SessionsController
  - Add additional test for login without account_id or alias
  - Support for password resets
  - Support changing passwords
  - Use our devise base class
  - Add confirmable support
  - Add email and confirmable
  - Custom mailer that handles jwt tokens
  - Allow email in request

- core: 
  - Configure sane SMTP defaults
  - Clarify that we override the token
  - We need the actionmailer railtie for SMTP

- services:
  - All services now depend on ActionMailer

# How does the implementation addresses the problem

This provides the reset password functionality. the main idea is that when the password reset flow starts, we create a JWT token that includes the Tenant id, the account id and the reset_password_token. This JWT is encoded server side and should not be tampered by the FE in any manner. Tampering would render the token invalid for the server.
The FE should then send the request with the token in the link, combined with the password and password_confirmation which would be used to reset the users password