core: only load mq/storage if explitly enabled
# Problem
By default ros-cli will write out config for mq and storage and the `infra.rb` loader doesn't care if they have been enabled but only if there is configuration.

As a consequence, ros will try to load mq even if not explicitly enabled causing things to blow up as `aws-sdk-sqs` is not included in core.

# Solution
Check if we are explicitly enabled before loading mq.

# Background
Due to the way we build the containers, the `aws-sdk-sqs` gem was being included everywhere thus hiding the problem. The workaround mentioned in #150 is now no longer needed.