[PW-3145] Use RequestStore to store user credential
# Refer to the issue id if any. Include the link to the issue. E.g:
[Use RequestStore to store user credential instead of class attributes](https://perxtechnologies.atlassian.net/browse/PW-3145)

# Describe the changes made. What does this PR changes that might be critical. If any critical decisions have been made, make sure you explain the rationale for these decisions.

- Use RequestStore to store user credential Instead class attributes, so that the credential is properly binding the incoming request.

# How does the implementation addresses the problem

After merging this, we store the user credentials in a request store. This way the credentials will be scoped per request and we should not have any issue where the in memory credentials get changed by another incoming request. Everytime we change the tenant, depending on the request timings, we were running the risk of changing the credentials accidentally.